Healthcare_project/care-connect/Project/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.3.2
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
django-rest-auth==0.9.5
django-allauth==0.61.1
drf-spectacular==0.27.1
django-debug-toolbar==4.3.0
django-storages==1.14.2
django-redis==5.4.0
celery==5.3.6
gunicorn==21.2.0
whitenoise==6.6.0
django-health-check==3.17.0
django-ratelimit==4.1.0 